+++ tasks/testnet.js
@@ -3,9 +3,8 @@
 let runDev = require(`./runner.js`)
 let config = require(`../app/src/config.js`)
 const path = require(`path`)
-const childProcess = require(`child_process`)
 const userHome = require(`user-home`)
-const fs = require(`fs-extra`)
+const { startLocalNode } = require(`./gaia.js`)
 
 async function main() {
   const network = process.argv[2] || config.default_network
@@ -14,10 +13,11 @@ async function main() {
 
   if (network === `local-testnet`) {
     extendedEnv = {
-      LCD_URL: `http://localhost:9070`,
+      LCD_URL: `https://localhost:9070`,
       RPC_URL: `http://localhost:26657`
     }
-    startLocalNode()
+    const TESTNET_NODE_FOLDER = path.join(userHome, `.gaiad-testnet`)
+    startLocalNode(TESTNET_NODE_FOLDER)
   }
 
   // run Voyager in a development environment
@@ -33,34 +33,3 @@ async function main() {
 main().catch(function(err) {
   console.error(`Starting the application failed`, err)
 })
-
-async function startLocalNode() {
-  const TESTNET_NODE_FOLDER = path.join(userHome, `.gaiad-testnet`)
-  const osFolderName = {
-    win32: `windows_amd64`,
-    darwin: `darwin_amd64`,
-    linux: `linux_amd64`
-  }[process.platform]
-  const gaiadFileName = process.platform === `win32` ? `gaiad.exe` : `gaiad`
-  const gaiadPath = path.join(`./builds/Gaia`, osFolderName, gaiadFileName)
-
-  console.log(`Starting local node`)
-  let child = childProcess.spawn(gaiadPath, [
-    `start`,
-    `--home`,
-    TESTNET_NODE_FOLDER
-  ])
-
-  // log output
-  const PROCESS_LOG = path.join(TESTNET_NODE_FOLDER, `process.log`)
-  const log = fs.createWriteStream(PROCESS_LOG, { flags: `a` })
-  console.log(`Find the node process logs at: ${PROCESS_LOG}`)
-  child.stdout.pipe(log)
-  child.stderr.pipe(process.stderr)
-
-  // handle node crashed
-  child.on(`exit`, () => {
-    console.error(`Local node crashed`)
-    process.exit()
-  })
-}