+++ test/unit/specs/App.spec.js
@@ -53,15 +53,18 @@ describe(`App without analytics`, () => {
     electron.remote.getGlobal = () => ({
       env: { NODE_ENV: `test` },
       mocked: true,
-      node_lcd: `https://awesomenode.de:12345`
+      node_lcd: `https://awesomenode.de:12345`,
+      development: false,
+      lcd_port_prod: `8080`
     })
     let Node = require(`renderer/connectors/node.js`)
     require(`renderer/main.js`)
-    expect(Node.mock.calls[0].slice(1)).toEqual([
-      `http://localhost:8080`,
+    expect(Node).toHaveBeenCalledWith(
+      expect.any(Function),
+      `https://localhost:8080`, // axios or axios proxy
       `https://awesomenode.de:12345`,
       true
-    ])
+    )
     jest.resetModules()
   })