+++ test/unit/specs/lcdClient.spec.js
@@ -286,16 +286,46 @@ describe(`LCD Client`, () => {
         ])
       })
 
-      it(`queries for a delegation summary for a delegator`, async () => {
+      it(`queries for a delegations of a delegator`, async () => {
         axios.mockReturnValue({})
-        await client.getDelegator(`abc`)
+        await client.getDelegations(`abc`)
 
         expect(axios.mock.calls).toEqual([
           [
             {
               data: undefined,
               method: `GET`,
-              url: `http://remotehost/stake/delegators/abc`
+              url: `http://remotehost/stake/delegators/abc/delegations`
+            }
+          ]
+        ])
+      })
+
+      it(`queries for a undelegations of a delegator`, async () => {
+        axios.mockReturnValue({})
+        await client.getUndelegations(`abc`)
+
+        expect(axios.mock.calls).toEqual([
+          [
+            {
+              data: undefined,
+              method: `GET`,
+              url: `http://remotehost/stake/delegators/abc/unbonding_delegations`
+            }
+          ]
+        ])
+      })
+
+      it(`queries for a redelegations of a delegator`, async () => {
+        axios.mockReturnValue({})
+        await client.getRedelegations(`abc`)
+
+        expect(axios.mock.calls).toEqual([
+          [
+            {
+              data: undefined,
+              method: `GET`,
+              url: `http://remotehost/stake/delegators/abc/redelegations`
             }
           ]
         ])
@@ -577,7 +607,7 @@ describe(`LCD Client`, () => {
       })
 
       it(`queries for governance txs`, async () => {
-        axios.mockReturnValue({})
+        axios.mockReturnValue({ data: [] })
         await client.getGovernanceTxs(lcdClientMock.addresses[0])
 
         expect(axios.mock.calls).toEqual([
@@ -600,6 +630,53 @@ describe(`LCD Client`, () => {
             }
           ]
         ])
+      })
+
+      describe(`Parameters`, function() {
+        it(`queries for governance deposit parameters`, async () => {
+          axios.mockReturnValue({})
+          await client.getGovDepositParameters()
+
+          expect(axios.mock.calls).toEqual([
+            [
+              {
+                data: undefined,
+                method: `GET`,
+                url: `http://remotehost/gov/parameters/deposit`
+              }
+            ]
+          ])
+        })
+
+        it(`queries for governance tallying parameters`, async () => {
+          axios.mockReturnValue({})
+          await client.getGovTallyingParameters()
+
+          expect(axios.mock.calls).toEqual([
+            [
+              {
+                data: undefined,
+                method: `GET`,
+                url: `http://remotehost/gov/parameters/tallying`
+              }
+            ]
+          ])
+        })
+
+        it(`queries for governance voting parameters`, async () => {
+          axios.mockReturnValue({})
+          await client.getGovVotingParameters()
+
+          expect(axios.mock.calls).toEqual([
+            [
+              {
+                data: undefined,
+                method: `GET`,
+                url: `http://remotehost/gov/parameters/voting`
+              }
+            ]
+          ])
+        })
       })
     })
 
@@ -705,7 +782,7 @@ describe(`LCD Client`, () => {
           {
             data: undefined,
             method: `GET`,
-            url: `http://remotehost/slashing/signing_info/pubKey`
+            url: `http://remotehost/slashing/validators/pubKey/signing_info`
           }
         ]
       ])